got up force move working for going up board. added new player control on player board to cope with force move.

+++ CheckersDA/ViewModels/MainMenu.cs
@@ -33,7 +33,7 @@ namespace CheckersDA.ViewModels
         {
             get
             {
-                switchMenuSel = Convert.ToInt32(menuSel);
+                switchMenuSel = Convert.ToInt16(menuSel);
                 return switchMenuSel;
             }
             set
